as.data.frame {spatPomp}	R Documentation

Coerce to data frame

Description

spatPomp objects can be recast as data frames.

Usage

## S3 method for class 'spatPomp'
as.data.frame(x, ...)

Arguments

x
a spatPomp object.

...
additional arguments to be passed to or from methods.

Details

When object is a simple ‘spatPomp’ object, as(object,"data.frame") or as.data.frame(object) results in a data frame with the times, units, observables, states (if known), and interpolated covariates (if any).

Value

A ‘data.frame’ with columns for time, spatial unit and observations.

[Package spatPomp version 0.29.0.1 Index]