bm {spatPomp}	R Documentation

Brownian motion spatPomp simulator

Description

Generate a class ‘spatPomp’ object representing a U-dimensional Brownian motion with spatial correlation decaying geometrically with distance around a circle. The model is defined in continuous time though in this case an Euler approximation is exact at the evaluation times.

Usage

bm(U = 5, N = 100, delta_t = 0.1)

Arguments

U
A length-one numeric signifying dimension of the process.

N
A length-one numeric signifying the number of observation time steps to evolve the process.

delta_t
Process simulations are performed every delta_t time units whereas observations occur every one time unit

Value

An object of class ‘spatPomp’ representing a simulation from a U-dimensional Brownian motion

Author(s)

Edward L. Ionides

See Also

Other spatPomp examples: he10(), lorenz(), measles()

Examples

# Complete examples are provided in the package tests
## Not run: 
b <- bm(U=4, N=20)
# See all the model specifications of the object
spy(b)
# Examples of methodologies applied to this model
# are provided in the tests directory

## End(Not run)

[Package spatPomp version 0.30.0.0 Index]