city_data_UK {spatPomp}	R Documentation

City data in the United Kingdom

Description

Population and birth information about cities in England and Wales during the measles pre-vaccine era.

Details

Data includes births and population at bi-weekly observations from 40 cities and towns.

Value

a ‘data.frame’ of the 40 largest cities and towns in the UK and Wales, their latitude, longitude and mean population during the measles pre-vaccine period.

References

Dalziel, Benjamin D. et al. (2016) Persistent chaos of measles epidemics in the prevaccination United States caused by a small change in seasonal transmission patterns. PLoS computational biology, 12(2), e1004655. DOI: 10.5061/dryad.r4q34

See Also

Other datasets: measlesUK

[Package spatPomp version 0.29.0.1 Index]