gbm {spatPomp}	R Documentation

Geometric Brownian motion spatPomp simulator

Description

Generate a spatPomp object representing a U-dimensional geometric Brownian motion with spatial correlation decaying geometrically with distance around a circle. The model is defined in continuous time, but an Euler approximation is used for this numerical implementation.

Usage

gbm(U = 5, N = 100, delta_t = 0.1, IVP_values = 1, delta_obs = 1)

Arguments

U
A length-one numeric signifying dimension of the process.

N
A length-one numeric signifying the number of time steps to evolve the process.

delta_t
process simulations are performed every delta_t time units

IVP_values
initial value parameters for the latent states

delta_obs
observations occur every delta_obs time units

Value

An object of class ‘spatPomp’ representing a simulation from a U-dimensional geometric Brownian motion

Examples

g <- gbm(U=4, N=20)
# See all the model specifications of the object
spy(g)

[Package spatPomp version 0.29.0.1 Index]