lorenz {spatPomp}	R Documentation

Geometric Brownian motion spatPomp simulator

Description

Generate a spatPomp object representing a U-dimensional geometric Brownian motion with spatial correlation decaying geometrically with distance around a circle. The model is defined in continuous time, but an Euler approximation is used for this numerical implementation.

Generate a spatPomp object representing a U-dimensional stochastic Lorenz '96 process with N measurements made at times t_n = n * delta_obs, simulated using an Euler method with time increment delta_t.

Usage

gbm(U = 5, N = 100, delta_t = 0.1, IVP_values = 1, delta_obs = 1)

lorenz(
  U = 5,
  N = 100,
  delta_t = 0.01,
  delta_obs = 0.5,
  regular_params = c(F = 8, sigma = 1, tau = 1)
)

Arguments

U
A length-one numeric signifying the number of spatial units for the process.

N
A length-one numeric signifying the number of observations.

delta_t
A length-one numeric giving the Euler time step for the numerical solution.

IVP_values
initial value parameters for the latent states

delta_obs
A length-one numeric giving the time between observations.

regular_params
A named numeric vector containing the values of the F, sigma and tau parameters. F=8 is a common value that causes chaotic behavior.

Value

An object of class ‘spatPomp’ representing a simulation from a U-dimensional geometric Brownian motion

An object of class ‘spatPomp’ representing a simulation from a U-dimensional Lorenz 96 model

Author(s)

Kidus Asfaw

Edward L. Ionides

References

Asfaw, K. T. (2021). Simulation-based Inference for Partially Observed Markov Process Models with Spatial Coupling. University of Michigan Doctoral dissertation. doi:10.7302/2751

Lorenz, E. N. (1996) Predictability: A problem partly solved. Proceedings of the seminar on predictability

Ionides, E. L., Asfaw, K., Park, J., and King, A. A. (2021). Bagged filters for partially observed interacting systems. Journal of the American Statistical Association, doi:10.1080/01621459.2021.1974867

See Also

Other spatPomp examples: bm(), he10(), measles()

Other spatPomp examples: bm(), he10(), measles()

Examples

# Complete examples are provided in the package tests
## Not run: 
g <- gbm(U=4, N=20)
# See all the model specifications of the object
spy(g)

## End(Not run)

# Complete examples are provided in the package tests
## Not run: 
l <- lorenz(U=5, N=100, delta_t=0.01, delta_obs=1)
# See all the model specifications of the object
spy(l)

## End(Not run)

[Package spatPomp version 0.30.0.0 Index]