mcap {spatPomp}	R Documentation

Monte Carlo adjusted profile

Description

Given a collection of points maximizing the likelihood over a range of fixed values of a focal parameter, this function constructs a profile likelihood confidence interval accommodating both Monte Carlo error in the profile and statistical uncertainty present in the likelihood function.

Usage

mcap(lp, parameter, confidence = 0.95, lambda = 1, Ngrid = 1000)

Arguments

lp
a vector of profile likelihood evaluations.

parameter
the corresponding values of the focal parameter.

confidence
the required level of the confidence interval.

lambda
the loess parameter used to smooth the profile.

Ngrid
the number of points to evaluate the smoothed profile.

Value

mcap() returns a list including the smoothed profile, a quadratic approximation, and the constructed confidence interval.

Author(s)

Edward L. Ionides

[Package spatPomp version 0.29.0.1 Index]