measlesUK {spatPomp}	R Documentation

Measles in the United Kingdom

Description

Measles case data from various cities and towns in England and Wales during the pre-vaccine era.

Details

Data includes bi-weekly case counts as well as births and population from 40 cities and towns.

Value

a ‘data.frame’ of the 40 largest cities and towns in the UK and Wales, their latitude, longitude and bi-weekly measles case counts, population and birthrates.

References

Dalziel, Benjamin D. et al. (2016) Persistent chaos of measles epidemics in the prevaccination United States caused by a small change in seasonal transmission patterns. PLoS computational biology, 12(2), e1004655. doi:10.5061/dryad.r4q34

See Also

Other datasets: city_data_UK, he10()

[Package spatPomp version 0.30.0.0 Index]