plot {spatPomp}	R Documentation

Plotting spatPomp data

Description

Visualize spatPomp data

Diagnostic plot for igirf()

Visualize spatPomp data

Usage

## S4 method for signature 'igirfd_spatPomp'
plot(x, params = names(coef(x)), ncol = 3)

## S4 method for signature 'spatPomp'
plot(x, type = c("l", "h"), log = F, ...)

Arguments

x
a spatPomp object

params
the names of the parameters for which the user would like to see a trace plot

ncol
the number of columns in the grid plot

type
for visualizing an object of class spatPomp, the user can obtain a grid of line plots by default ('l') or a heat map by supplying argument 'h'.

log
should the data be log-transformed before plotting? This helps in contexts where there are spikes that could take away attention from the dynamics illustrated by the rest of the data.

...
for visualizing an object of class spatPomp, the user can add arguments like nrow to specify the number of rows in the grid.

Value

a ggplot facet plot of class ‘gg’ and ‘ggplot’ visualizing the convergence record of running igirf() with respect to the likelihood and the parameters of the model.

a ggplot plot of class ‘gg’ and ‘ggplot’ visualizing the time series data over multiple spatial units via a tile-plot.

[Package spatPomp version 0.29.0.1 Index]