runit_measure {spatPomp}	R Documentation

runit_measure

Description

runit_measure simulates a unit's observation given the entire state

Usage

## S4 method for signature 'spatPomp'
runit_measure(object, x, unit, time, params, log = FALSE)

Arguments

object
An object of class spatPomp

x
A state vector for all units

unit
The unit for which to simulate an observation

time
The time for which to simulate an observation

params
parameters to use to simulate an observation

log
logical; should the density be returned on log scale?

Value

A matrix with the simulated observation corresponding to state x and unit unit with parameter set params.

Examples

b <- bm(U=3)
s <- states(b)[,1,drop=FALSE]
rownames(s) -> rn
dim(s) <- c(3,1,1)
dimnames(s) <- list(variable=rn, rep=NULL)
p <- coef(b); names(p) -> rnp
dim(p) <- c(length(p),1); dimnames(p) <- list(param=rnp)
o <- obs(b)[,1,drop=FALSE]
runit_measure(b, x=s, unit=2, time=1, params=p)

[Package spatPomp version 0.29.0.1 Index]