simulate {spatPomp}	R Documentation

Simulation of a spatiotemporal partially-observed Markov process

Description

simulate generates simulations of the latent and measurement processes.

Usage

## S4 method for signature 'spatPomp'
simulate(
  object,
  nsim = 1,
  seed = NULL,
  format = c("spatPomps", "data.frame"),
  include.data = FALSE,
  ...
)

Arguments

object
optional; if present, it should be a data frame or a ‘pomp’ object.

nsim
number of simulations.

seed
optional; if set, the pseudorandom number generator (RNG) will be initialized with seed. the random seed to use. The RNG will be restored to its original state afterward.

format
the format of the simulated results. If the argument is set to 'spatPomps', the default behavior, then the output is a list of spatPomp objects. Options are 'spatPomps' and 'data.frame'.

include.data
if TRUE, the original data and covariates (if any) are included (with .id = "data"). This option is ignored unless format = "data.frame".

...
additional arguments supply new or modify existing model characteristics or components. See pomp for a full list of recognized arguments.

When named arguments not recognized by pomp are provided, these are made available to all basic components via the so-called userdata facility. This allows the user to pass information to the basic components outside of the usual routes of covariates (covar) and model parameters (params). See userdata for information on how to use this facility.

Value

if format='spatPomps' and nsim=1 an object of class ‘spatPomp’ representing a simulation from the model in object is returned. If format='spatPomps' and nsim>1 a list of class ‘spatPomp’ objects is returned. If format='data.frame' then a class ‘data.frame’ object is returned.

Author(s)

Kidus Asfaw

References

Asfaw, K., Park, J., Ho, A., King, A. A., and Ionides, E. L. (2020) Partially observed Markov processes with spatial structure via the R package spatPomp. ArXiv: 2101.01157. doi:10.48550/arXiv.2101.01157

Examples

# Complete examples are provided in the package tests
## Not run: 
# Get a spatPomp object
b <- bm(U=2, N=5)
# Get 2 simulations from same model as data.frame
sims <- simulate(b, nsim=2, format='data.frame')

## End(Not run)

[Package spatPomp version 0.30.0.0 Index]