spatPomp-class {spatPomp}	R Documentation

An S4 class to represent a spatiotemporal POMP model and data.

Description

An S4 class to represent a spatiotemporal POMP model and data.

Slots

unit_names

A vector containing the spatial units of a spatiotemporal POMP.

unit_statenames

A vector containing the state names such that appending the unit indices to the unit statenames will result in the each unit's corresponding states.

unit_obsnames

A vector of observation types for a spatial unit.

eunit_measure

A pomp_fun representing the expected measurement for each spatial unit given its states.

dunit_measure

A pomp_fun representing the unit measurement density for each spatial unit.

runit_measure

A pomp_fun representing the unit observation simulator.

[Package spatPomp version 0.30.0.0 Index]