spatPomp-package {spatPomp}	R Documentation

Inference for SpatPOMPs (Spatiotemporal Partially Observed Markov Processes)

Description

The spatPomp package provides facilities for inference on panel data using spatiotemporal partially-observed Markov process (SpatPOMP) models. To do so, it relies on and extends a number of facilities that the pomp package provides for inference on time series data using partially-observed Markov process (POMP) models.

The spatPomp package concerns models consisting of a collection of interacting units. The methods in spatPomp may be applicable whether or not these units correspond to spatial locations.

Data analysis using spatPomp

The first step in using spatPomp is to encode one's model(s) and data in objects of class spatPomp. This can be done via a call to the spatPomp constructor function.

Extending the pomp platform for developing inference tools

spatPomp extends to panel data the general interface to the components of POMP models provided by pomp. In doing so, it contributes to the goal of the pomp project of facilitating the development of new algorithms in an environment where they can be tested and compared on a growing body of models and datasets.

Documentation

spatPomp is described by Asfaw et al. (2020)

License

spatPomp is provided under the MIT License.

Author(s)

Kidus Asfaw, Joonha Park, Allister Ho, Edward Ionides, Aaron King

References

Asfaw, K. et al. (2020) Statistical inference for spatiotemporal partially observed Markov processes via the R package spatPomp. Manuscript in preparation.

See Also

pomp package

[Package spatPomp version 0.29.0.1 Index]