unit_names {spatPomp}	R Documentation

Unit names of a spatiotemporal model

Description

unit_names outputs the contents of the unit_names slot of a spatPomp object. The order in which the units appear in the output vector determines the order in which latent states and observations for the spatial units are stored.

Usage

## S4 method for signature 'spatPomp'
unit_names(x)

Arguments

x
a spatPomp object

Value

A character vector with the unit names used to create the ‘spatPomp’ object.

[Package spatPomp version 0.29.0.1 Index]