vec_dmeasure {spatPomp}	R Documentation

Vector of measurement densities

Description

Evaluate the unit measurement model density function for each unit. This method is used primarily as part of likelihood evaluation and parameter inference algorithms.

Usage

## S4 method for signature 'spatPomp'
vec_dmeasure(object, y, x, units, times, params, log = FALSE, ...)

Arguments

object
a spatPomp object

y
numeric; measurements whose densities given the latent states are evaluated

x
numeric; state at which conditional measurement densities are evaluated

units
numeric; units at which measurement densities are evaluated

times
numeric; time at which measurement densities are evaluated

params
numeric; parameter set at which measurement densities is evaluated

log
logical; should the outputted measurement densities be on log scale?

...
additional parameters will be ignored

Value

An array of dimension length(unit_names(object)) by dim(x)[2] by dim(x)[3] representing each unit's measurement density assessed for each replicate in x for each observation time.

[Package spatPomp version 0.29.0.1 Index]