lorenz {spatPomp}	R Documentation

Lorenz '96 spatPomp simulator

Description

Generate a spatPomp object representing a U-dimensional stochastic Lorenz '96 process with N measurements made at times t_n = n * delta_obs, simulated using an Euler method with time increment delta_t.

Usage

lorenz(
  U = 5,
  N = 100,
  delta_t = 0.01,
  delta_obs = 0.5,
  regular_params = c(F = 8, sigma = 1, tau = 1)
)

Arguments

U
A length-one numeric signifying the number of spatial units for the process.

N
A length-one numeric signifying the number of observations.

delta_t
A length-one numeric giving the Euler time step for the numerical solution.

delta_obs
A length-one numeric giving the time between observations.

regular_params
A named numeric vector containing the values of the F, sigma and tau parameters. F=8 is a common value that causes chaotic behavior.

Value

An object of class ‘spatPomp’ representing a simulation from a U-dimensional Lorenz 96 model

References

Lorenz, E. N. (1996) Predictability: A problem partly solved. Proceedings of the seminar on predictability

Examples

l <- lorenz(U=5, N=100, delta_t=0.01, delta_obs=1)
# See all the model specifications of the object
spy(l)

[Package spatPomp version 0.29.0.1 Index]